refactor: 로깅 분리
